Struct uni_app::AppConfig

pub struct AppConfig {
    pub title: String,
    pub size: (u32, u32),
    pub vsync: bool,
    pub headless: bool,
    pub fullscreen: bool,
    pub resizable: bool,
    pub show_cursor: bool,
    pub intercept_close_request: bool,
}

game window configuration

Fields
title: String

the window title (only visible on native target)

size: (u32, u32)

the window/canvas size in pixels

vsync: bool

sync frames with screen frequency (can only be disabled on native target)

headless: bool

start the program without actually creating a window, for test purposes

fullscreen: bool

start in full screen (native target only)

resizable: bool

whether user can resize the window (native target only)

show_cursor: bool

whether the mouse cursor is visible while in the window

intercept_close_request: bool

whether clicking on the window close button exits the program or sends a CloseRequested event

Implementations

impl AppConfig

pub fn new<T: Into<String>>(title: T, size: (u32, u32)) -> AppConfig